% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m750_training_resamples_fitted}
\alias{m750_training_resamples_fitted}
\title{Time Series Cross Validation Resample Predictions (Results) from the M750 Data (Training Set)}
\format{
A Modeltime Table that has been fitted to resamples with predictions in the \code{.resample_results} column
}
\usage{
m750_training_resamples_fitted
}
\description{
Time Series Cross Validation Resample Predictions (Results) from the M750 Data (Training Set)
}
\details{
\if{html}{\out{<div class="sourceCode {r eval = FALSE}#'">}}\preformatted{m750_training_resamples_fitted <- m750_models \%>\%
    modeltime_fit_resamples(
        resamples = m750_training_resamples,
        control   = control_resamples(verbose = T)
    )
}\if{html}{\out{</div>}}
}
\examples{

m750_training_resamples_fitted



}
\seealso{
\itemize{
\item \link[modeltime:m750_models]{modeltime::m750_models}
\item \link[modeltime:m750_training_resamples]{modeltime::m750_training_resamples}
}
}
\keyword{datasets}
