% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-timetk.R
\name{timetk}
\alias{timetk}
\alias{time_series_cv}
\alias{time_series_split}
\alias{plot_time_series_cv_plan}
\alias{tk_time_series_cv_plan}
\title{Time Series Resampling Helpers}
\description{
\itemize{
\item \code{\link[timetk:time_series_cv]{timetk::time_series_cv()}}: Creates resample sets using time series cross validation
\item \code{\link[timetk:time_series_split]{timetk::time_series_split()}}: Makes an initial time series split
\item \code{\link[timetk:plot_time_series_cv_plan]{timetk::plot_time_series_cv_plan()}}: Plots a cross validation plan
\item \code{\link[timetk:tk_time_series_cv_plan]{timetk::tk_time_series_cv_plan()}}: Unnests a cross validation plan
}
}
\keyword{internal}
